Rename join_common -> JoinUtils
